requirements.txt
# Core Dependencies
praw==7.7.1                
openai==0.28.1
python-dotenv==1.0.1      
requests==2.32.3           
urllib3==2.2.1             
python-dateutil==2.9.0    